feat: introduce `FoldingScheme` and unify `sangria` and `protostar`
