--- plonkish_backend/src/backend/hyperplonk/folding.rs
@@ -1,10 +1,0 @@
-mod protostar;
-mod sangria;
-
-pub use protostar::{
-    Protostar, ProtostarProverParam, ProtostarProverState, ProtostarVerifierParam,
-};
-pub use sangria::{Sangria, SangriaProverParam, SangriaProverState, SangriaVerifierParam};
-
-#[cfg(feature = "frontend-halo2")]
-pub use protostar::verifier::halo2::*;
