+++ plonkish_backend/src/folding/protostar/hyperplonk/ivc.rs
@@ -1,0 +1,3 @@
+// TODO:
+// #[cfg(feature = "frontend-halo2")]
+// mod halo2;
