+++ plonkish_backend/src/lib.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::op_ref)]
 
 pub mod backend;
+pub mod folding;
 pub mod frontend;
 pub mod pcs;
 pub mod piop;
