+++ plonkish_backend/src/pcs/multilinear/ipa.rs
@@ -1,9 +1,9 @@
 use crate::{
     pcs::{
         multilinear::{additive, err_too_many_variates, validate_input},
-        AdditiveCommitment, Evaluation, Point, Polynomial, PolynomialCommitmentScheme,
+        AdditiveCommitment, Evaluation, Point, PolynomialCommitmentScheme,
     },
-    poly::multilinear::MultilinearPolynomial,
+    poly::{multilinear::MultilinearPolynomial, Polynomial},
     util::{
         arithmetic::{
             inner_product, variable_base_msm, Curve, CurveAffine, CurveExt, Field, Group,
