+++ plonkish_backend/src/pcs/multilinear/kzg.rs
@@ -1,9 +1,9 @@
 use crate::{
     pcs::{
         multilinear::{additive, err_too_many_variates, validate_input},
-        AdditiveCommitment, Evaluation, Point, Polynomial, PolynomialCommitmentScheme,
+        AdditiveCommitment, Evaluation, Point, PolynomialCommitmentScheme,
     },
-    poly::multilinear::MultilinearPolynomial,
+    poly::{multilinear::MultilinearPolynomial, Polynomial},
     util::{
         arithmetic::{
             div_ceil, fixed_base_msm, variable_base_msm, window_size, window_table, Curve,
