+++ plonkish_backend/src/util/expression/relaxed.rs
@@ -1,6 +1,6 @@
 use crate::util::{
     arithmetic::PrimeField,
-    expression::{Expression, Query},
+    expression::{CommonPolynomial, Expression, Query},
     BitIndex, Itertools,
 };
 use std::{
@@ -14,7 +14,24 @@ pub(crate) struct PolynomialSet {
     pub(crate) folding: BTreeSet<usize>,
 }
 
-pub(crate) fn cross_term_expressions<F: PrimeField + Ord>(
+#[derive(Clone, PartialEq, PartialOrd, Eq, Ord)]
+enum ExpressionPolynomial {
+    CommonPolynomial(CommonPolynomial),
+    Polynomial(Query),
+}
+
+impl<F> From<ExpressionPolynomial> for Expression<F> {
+    fn from(poly_foldee: ExpressionPolynomial) -> Self {
+        match poly_foldee {
+            ExpressionPolynomial::CommonPolynomial(common_poly) => {
+                Expression::CommonPolynomial(common_poly)
+            }
+            ExpressionPolynomial::Polynomial(query) => Expression::Polynomial(query),
+        }
+    }
+}
+
+pub(crate) fn cross_term_expressions<F: PrimeField>(
     poly_set: &PolynomialSet,
     products: &[Product<F>],
     num_challenges: usize,
@@ -28,17 +45,20 @@ pub(crate) fn cross_term_expressions<F: PrimeField + Ord>(
     products
         .iter()
         .fold(
-            vec![BTreeMap::<Vec<Expression<F>>, Expression<F>>::new(); num_ts],
+            vec![BTreeMap::<Vec<_>, Expression<F>>::new(); num_ts],
             |mut scalars, product| {
-                let (common_scalar, common_expr) = product.preprocess.evaluate(
+                let (common_scalar, common_poly) = product.preprocess.evaluate(
                     &|constant| (constant, Vec::new()),
-                    &|common_poly| (F::ONE, vec![Expression::CommonPolynomial(common_poly)]),
-                    &|query| {
-                        let poly = preprocess_poly_indices[&query.poly()];
+                    &|common_poly| {
                         (
                             F::ONE,
-                            vec![Expression::Polynomial(Query::new(poly, query.rotation()))],
+                            vec![ExpressionPolynomial::CommonPolynomial(common_poly)],
                         )
+                    },
+                    &|query| {
+                        let poly = preprocess_poly_indices[&query.poly()];
+                        let query = Query::new(poly, query.rotation());
+                        (F::ONE, vec![ExpressionPolynomial::Polynomial(query)])
                     },
                     &|_| unreachable!(),
                     &|(scalar, expr)| (-scalar, expr),
@@ -49,13 +69,13 @@ pub(crate) fn cross_term_expressions<F: PrimeField + Ord>(
                     &|(lhs, expr), rhs| (lhs * rhs, expr),
                 );
                 for idx in 1usize..(1 << folding_degree) - 1 {
-                    let (scalar, mut exprs) = iter::empty()
+                    let (scalar, mut polys) = iter::empty()
                         .chain(iter::repeat(None).take(folding_degree - product.folding_degree()))
                         .chain(product.foldees.iter().map(Some))
                         .enumerate()
                         .fold(
-                            (Expression::Constant(common_scalar), common_expr.clone()),
-                            |(mut scalar, mut exprs), (nth, foldee)| {
+                            (Expression::Constant(common_scalar), common_poly.clone()),
+                            |(mut scalar, mut polys), (nth, foldee)| {
                                 let (poly_offset, challenge_offset) = if idx.nth_bit(nth) {
                                     (
                                         preprocess_poly_indices.len() + folding_poly_indices.len(),
@@ -77,16 +97,16 @@ pub(crate) fn cross_term_expressions<F: PrimeField + Ord>(
                                         let poly =
                                             poly_offset + folding_poly_indices[&query.poly()];
                                         let query = Query::new(poly, query.rotation());
-                                        exprs.push(Expression::Polynomial(query));
+                                        polys.push(ExpressionPolynomial::Polynomial(query));
                                     }
                                     _ => unreachable!(),
                                 }
-                                (scalar, exprs)
+                                (scalar, polys)
                             },
                         );
-                    exprs.sort_unstable();
+                    polys.sort_unstable();
                     scalars[idx.count_ones() as usize - 1]
-                        .entry(exprs)
+                        .entry(polys)
                         .and_modify(|value| *value = value as &Expression<_> + &scalar)
                         .or_insert(scalar);
                 }
@@ -97,7 +117,13 @@ pub(crate) fn cross_term_expressions<F: PrimeField + Ord>(
         .map(|exprs| {
             exprs
                 .into_iter()
-                .map(|(polys, scalar)| polys.into_iter().product::<Expression<_>>() * scalar)
+                .map(|(polys, scalar)| {
+                    polys
+                        .into_iter()
+                        .map_into::<Expression<F>>()
+                        .product::<Expression<_>>()
+                        * scalar
+                })
                 .sum::<Expression<_>>()
         })
         .collect_vec()
